Use the existing World ID 4.0 route
